update: load data type

+++ factorization_machine/include/data_struct.h
@@ -2,7 +2,7 @@
 #define DATA_STRUCT_H_
 
 struct sparse_feature{
-    long int idx;
-    double val;
+    int idx;
+    int val;
 };
 #endif
