+++ factorization_machine/include/load_data.h
@@ -15,12 +15,12 @@ public:
     ~Load_Data();
     void load_data(const char *train_file, std::string split_tag, int rank, int nproc);
     std::string line;
-    float y, value;
+    int y, value;
     int nchar, index;
     std::vector<std::vector<sparse_feature> > fea_matrix;
     std::vector<sparse_feature> key_val;
     sparse_feature sf;
-    std::vector<float> label;
+    std::vector<int> label;
     long int loc_fea_dim = 0;
     long int glo_fea_dim = 0;
     long int loc_ins_num = 0;
